Removed old callpath generating functions.  Removed data structures associated with them for another 13MB heap savings on Miranda16k.
Removed max values associated with Functions.  This makes maintenance of
importers easier, and speeds them up a little too.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2957 a8e31b63-e6ca-409b-a01b-2eec1b506579

+++ tools/src/dms/src/edu/uoregon/tau/dms/dss/UserEvent.java
@@ -18,11 +18,11 @@ public class UserEvent implements Serializable, Comparable {
         return id;
     }
 
-    public void setMaxUserEventNumberValue(int maxUserEventNumberValue) {
+    public void setMaxUserEventNumberValue(double maxUserEventNumberValue) {
         this.maxUserEventNumberValue = maxUserEventNumberValue;
     }
 
-    public int getMaxUserEventNumberValue() {
+    public double getMaxUserEventNumberValue() {
         return maxUserEventNumberValue;
     }
 
@@ -102,7 +102,7 @@ public class UserEvent implements Serializable, Comparable {
     private Color specificColor;
     private boolean colorFlag = false;
     
-    private int maxUserEventNumberValue = 0;
+    private double maxUserEventNumberValue = 0;
     private double maxUserEventMinValue = 0;
     private double maxUserEventMaxValue = 0;
     private double maxUserEventMeanValue = 0;
